refactor(S3WritableByteChannel): Extract `S3TransferUtil` class responsible for downloading/uploading
