+++ src/main/java/software/amazon/nio/spi/s3/S3WritableByteChannel.java
@@ -2,12 +2,6 @@ package software.amazon.nio.spi.s3;
 
 import org.checkerframework.checker.nullness.qual.NonNull;
 import software.amazon.awssdk.services.s3.S3AsyncClient;
-import software.amazon.awssdk.services.s3.model.GetObjectRequest;
-import software.amazon.awssdk.services.s3.model.PutObjectRequest;
-import software.amazon.awssdk.transfer.s3.S3TransferManager;
-import software.amazon.awssdk.transfer.s3.model.DownloadFileRequest;
-import software.amazon.awssdk.transfer.s3.model.UploadFileRequest;
-
 import java.io.IOException;
 import java.nio.ByteBuffer;
 import java.nio.channels.SeekableByteChannel;
@@ -22,24 +16,20 @@ import java.util.Objects;
 import java.util.HashSet;
 import java.util.Set;
 import java.util.concurrent.ExecutionException;
-import java.util.concurrent.TimeUnit;
 import java.util.concurrent.TimeoutException;
 
 class S3WritableByteChannel implements WritableByteChannel {
-    private final S3AsyncClient client;
     private final S3Path path;
     private final Path tempFile;
     private final SeekableByteChannel channel;
-    private final Long timeout;
-    private final TimeUnit timeUnit;
+    private final S3TransferUtil s3TransferUtil;
 
     private boolean open;
 
-    S3WritableByteChannel(S3Path path, S3AsyncClient client, Set<? extends OpenOption> options, Long timeout, TimeUnit timeUnit) throws IOException {
+    S3WritableByteChannel(S3Path path, S3AsyncClient client, S3TransferUtil s3TransferUtil, Set<? extends OpenOption> options) throws IOException {
         Objects.requireNonNull(path);
         Objects.requireNonNull(client);
-        this.timeout = timeout;
-        this.timeUnit = timeUnit;
+        this.s3TransferUtil = s3TransferUtil;
 
         try {
             var fileSystemProvider = (S3FileSystemProvider) path.getFileSystem().provider();
@@ -54,23 +44,7 @@ class S3WritableByteChannel implements WritableByteChannel {
 
             tempFile = Files.createTempFile("aws-s3-nio-", ".tmp");
             if (exists) {
-                try (var s3TransferManager = S3TransferManager.builder().s3Client(client).build()) {
-                    var downloadCompletableFuture = s3TransferManager.downloadFile(
-                            DownloadFileRequest.builder()
-                                    .getObjectRequest(GetObjectRequest.builder()
-                                            .bucket(path.bucketName())
-                                            .key(path.getKey())
-                                            .build())
-                                    .destination(tempFile)
-                                    .build()
-                    ).completionFuture();
-
-                    if (timeout != null && timeUnit != null) {
-                        downloadCompletableFuture.get(timeout, timeUnit);
-                    } else {
-                        downloadCompletableFuture.join();
-                    }
-                }
+                s3TransferUtil.downloadToLocalFile(path, tempFile);
             }
 
             channel = Files.newByteChannel(this.tempFile, removeCreateNew(options));
@@ -81,7 +55,6 @@ class S3WritableByteChannel implements WritableByteChannel {
             throw new IOException("Could not open the path:" + path, e);
         }
 
-        this.client = client;
         this.path = path;
         this.open = true;
     }
@@ -106,29 +79,8 @@ class S3WritableByteChannel implements WritableByteChannel {
     public void close() throws IOException {
         channel.close();
 
-        try (var s3TransferManager = S3TransferManager.builder().s3Client(client).build()) {
-            var uploadCompletableFuture = s3TransferManager.uploadFile(
-                    UploadFileRequest.builder()
-                            .putObjectRequest(PutObjectRequest.builder()
-                                    .bucket(path.bucketName())
-                                    .key(path.getKey())
-                                    .contentType(Files.probeContentType(tempFile))
-                                    .build())
-                            .source(tempFile)
-                            .build()
-            ).completionFuture();
+        s3TransferUtil.uploadLocalFile(path, tempFile);
 
-            if (timeout != null && timeUnit != null) {
-                uploadCompletableFuture.get(timeout, timeUnit);
-            } else {
-                uploadCompletableFuture.join();
-            }
-        } catch (InterruptedException e) {
-            Thread.currentThread().interrupt();
-            throw new IOException("Could not write to path:" + path, e);
-        } catch (TimeoutException | ExecutionException e) {
-            throw new IOException("Could not write to path:" + path, e);
-        }
         open = false;
     }
 }
